requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14
opencv-python==4.8.1.78
wikipedia==1.4.0
psutil==5.9.6
pyautogui==0.9.54
pywhatkit==5.4
spacy==3.7.2
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.4
langchain-core>=0.1.7,<0.2
langchain-community==0.0.10
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
pywebview==5.2
numpy==1.26.2
matplotlib==3.8.2
pillow==10.1.0
aiofiles==23.2.1
pywin32==306; sys_platform == "win32"

